Rename 'extra-files' to 'downloads-plugin'
It's been two years since the original plugin had a revision...  but we need
most of these changes to be available on packagist.  (The raison d'etre is
to avoid extra configuration of custom-repos and root-projects.) Not optimistic
about the PR route per #2 -- so the other option is renaming.

Since we're renaming, it should be fairly thorough to ensure that we don't
accidentally introduces conflicts.