Change 'file' handling to fix re-download use-case and be more sensible
To understand this change, suppose you want to create a file `docs/hello.txt`

Before
------

Use this configuration:

```
{
  "url": "https://example.com/hello.txt",
  "path": "docs"
}
```

After
------

Use this configuration:

```
{
  "url": "https://example.com/hello.txt",
  "path": "docs/hello.txt"
}
```

Comments
--------

This plugin was built on top of `$downloadManager`, which is generally oriented
towards creating *folders* with files under them. This makes perfect sense when
extracting archives.

It feels wonky when downloading individual files -- because every file has to be
placed in a unique/separate folder.  As a user of this interface, I'd expect to
drop files into arbitrary places (e.g. putting `docs/hello.txt` and
`docs/world.txt` next to each other, even though they're not from the same source).
I'd also expect to be able to control the name of the file that's written.

This change modifies the contract so that `file` items feel more like
individual files.  It also fixes the re-download scenario for these files.